+++ sdk/deviceregistry/Azure.ResourceManager.DeviceRegistry/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Release History
 
+## 1.0.0-beta.3 (Unreleased)
+
+### Breaking Changes
+
+- Updated `Assets` and `AssetEndpointProfiles` resource types to align to Microsoft.DeviceRegistry OpenAPI spec version `2024-09-01-preview`.
+
+### Features Added
+
+- Added `SchemaRegistries`, `SchemaRegistries/Schemas` and `SchemaRegistries/Schemas/SchemaVersions` resource types from Microsoft.DeviceRegistry OpenAPI spec version `2024-09-01-preview`.
+- Added `BillingContainers` resource type from Microsoft.DeviceRegistry OpenAPI spec version `2024-09-01-preview`.
+
 ## 1.0.0-beta.2 (Unreleased)
 
 ### Features Added